requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
lxml>=4.9.0
html5lib>=1.1
openpyxl>=3.1.0
python-dotenv>=1.0.0
markdown>=3.4.0
plotly>=5.20.0

